+++ src/video/renderer.js
@@ -189,42 +189,6 @@ export default class Renderer {
     }
 
     /**
-     * Returns the 2D Context object of the given Canvas<br>
-     * Also configures anti-aliasing and blend modes based on constructor options.
-     * @param {HTMLCanvasElement} canvas
-     * @param {boolean} [transparent=true] - use false to disable transparency
-     * @returns {CanvasRenderingContext2D}
-     */
-    getContext2d(canvas, transparent) {
-        if (typeof canvas === "undefined" || canvas === null) {
-            throw new Error(
-                "You must pass a canvas element in order to create " +
-                "a 2d context"
-            );
-        }
-
-        if (typeof canvas.getContext === "undefined") {
-            throw new Error(
-                "Your browser does not support HTML5 canvas."
-            );
-        }
-
-        if (typeof transparent !== "boolean") {
-            transparent = true;
-        }
-
-        let _context = canvas.getContext("2d", {
-            "alpha" : transparent
-        });
-
-        if (!_context.canvas) {
-            _context.canvas = canvas;
-        }
-        this.setAntiAlias(_context, this.settings.antiAlias);
-        return _context;
-    }
-
-    /**
      * get the current fill & stroke style color.
      * @returns {Color} current global color
      */
@@ -273,11 +237,10 @@ export default class Renderer {
 
     /**
      * enable/disable image smoothing (scaling interpolation) for the given context
-     * @param {CanvasRenderingContext2D} context
      * @param {boolean} [enable=false]
      */
-    setAntiAlias(context, enable) {
-        this.renderTarget.setAntiAlias(context, enable);
+    setAntiAlias(enable) {
+        this.renderTarget.setAntiAlias(enable);
     }
 
     /**
@@ -340,25 +303,22 @@ export default class Renderer {
      * @param {HTMLImageElement|HTMLCanvasElement|OffscreenCanvas} src - the source image to be tinted
      * @param {Color|string} color - the color that will be used to tint the image
      * @param {string} [mode="multiply"] - the composition mode used to tint the image
-     * @returns {HTMLCanvasElement|OffscreenCanvas} a new canvas element representing the tinted image
+     * @returns {HTMLCanvasElement|OffscreenCanvas} a new canvas or offscreencanvas (if supported) element representing the tinted image
      */
-    tint(src, color, mode) {
-        let canvas = createCanvas(src.width, src.height, true);
-        let context = this.getContext2d(canvas);
-
-        context.save();
+    tint(src, color, mode = "multiply") {
+        const attributes = { context:"2d", offscreenCanvas: true, transparent: true, antiAlias: this.settings.antiAlias };
+        let canvasTexture = new CanvasRenderTarget(src.width, src.height, attributes);
+        let context = canvasTexture.context;
 
         context.fillStyle = color instanceof Color ? color.toRGB() : color;
         context.fillRect(0, 0, src.width, src.height);
 
-        context.globalCompositeOperation = mode || "multiply";
+        context.globalCompositeOperation = mode;
         context.drawImage(src, 0, 0);
         context.globalCompositeOperation = "destination-atop";
         context.drawImage(src, 0, 0);
 
-        context.restore();
-
-        return canvas;
+        return canvasTexture.canvas;
     }
 
     /**