+++ src/loader/parsers/fontface.js
@@ -1,3 +1,4 @@
+import { fontList } from "../cache.js";
 import { isDataUrl } from "../../utils/string.js";
 
 /**
@@ -13,6 +14,7 @@ import { isDataUrl } from "../../utils/string.js";
  * ]);
  */
 export function preloadFontFace(data, onload, onerror) {
+    const fontFaceSet = typeof globalThis.document !== "undefined" ? globalThis.document.fonts : undefined;
 
     if (isDataUrl(data.src) === true) {
         // make sure it in the `url(data:[<mediatype>][;base64],<data>)` format as expected by FontFace
@@ -21,23 +23,31 @@ export function preloadFontFace(data, onload, onerror) {
         }
     }
 
-    let font = new FontFace(data.name, data.src);
-
-    // loading promise
-    font.load().then(() => {
-        // apply the font after the font has finished downloading
-        globalThis.document.fonts.add(font);
-        globalThis.document.body.style.fontFamily = data.name;
-        if (typeof onload === "function") {
+    if (typeof fontFaceSet !== "undefined") {
+        // create a new font face
+        let font = new FontFace(data.name, data.src);
+        // loading promise
+        font.load().then(() => {
+            // add the font to the cache
+            fontList[data.name] = font;
+            // add the font to the document
+            fontFaceSet.add(font);
             // onloaded callback
-            onload();
-        }
-    }, () => {
-        if (typeof onerror === "function") {
+            if (typeof onload === "function") {
+                onload();
+            }
+        }, () => {
             // rejected
-            onerror(data.name);
+            if (typeof onerror === "function") {
+                onerror(data.name);
+            }
+        });
+
+    } else {
+        if (typeof onerror === "function") {
+            onerror(error);
         }
-    });
+    }
 
     return 1;
 }