+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Update to eslint 9.0
 
 ### Fixed 
-- Renderer: fix a regression when forcing WebGL1 rendering mode
+- Renderer: fix a regression when forcing WebGL1 rendering mode (leading to an exception)
+- Renderer: fix a regression when using the canvas rendering mode where antialias setting was not being applied on cached tinted elements
 
 ## [17.1.0] (melonJS 2) - _2024-03-29_