+++ CHANGELOG.md
@@ -1,14 +1,14 @@
 # Changelog
 
-## [17.2.0] (melonJS 2) - _2024-xx-xx_
+## [17.2.0] (melonJS 2) - _2024-04-22_
 
 ### Added
 - Audio: add missing optional id parameter to 3D / Spatial audio methods
 - Core: add platform detection if running as a standalone web app
-- Loader: properly return an error when attempting to load FontFace assets on non-browser platforms
 - Loader: add missing `unload` implementation for font assets
 
 ### Fixed
+- Loader: properly return an error when attempting to load FontFace assets on non-browser platforms
 - Renderer: fix a regression when a global canvas is available (e.g. webapp adapter for wechat)
 - Renderer: fix a regression when forcing WebGL1 rendering mode (leading to an exception)
 - Renderer: fix a regression when using the canvas rendering mode where antialias setting was not being applied on cached tinted elements